+++ SmartPackager/BasicPackMethods/PackDateTime.cs
@@ -8,20 +8,20 @@ namespace SmartPackager.BasicPackMethods
 
         public bool IsFixedSize => true;
 
-        public unsafe long PackUP(byte* destination, DateTime source)
+        public unsafe int PackUP(byte* destination, DateTime source)
         {
             *(long*)destination = source.Ticks;
             return sizeof(long);
         }
 
-        public unsafe long UnPack(byte* source, out DateTime destination)
+        public unsafe int UnPack(byte* source, out DateTime destination)
         {
             DateTime pt = new DateTime(*(long*)source);
             destination = pt;
             return sizeof(long);
         }
 
-        public long GetSize(DateTime source)
+        public int GetSize(DateTime source)
         {
             return sizeof(long);
         }
